started modularizing
