--- .env
@@ -1,23 +1,0 @@
-# .env
-# ANNOUNCEMENT_DAY = 0 for monday, 1 for tuesday, etc. 6 = sunday
-# ANNOUNCEMENT_HOUR = 0-24
-# ANNOUNCEMENT_MIN = 0-60
-
-DISCORD_TOKEN=
-DISCORD_GUILD=
-ANNOUNCEMENT_CHANNEL=
-ANNOUNCEMENT_DAY=
-ANNOUNCEMENT_HOUR=
-ANNOUNCEMENT_MIN=
-WELCOME_CHANNEL=
-WELCOME_ROLE=
-OPENDOTA_API_KEY=
-DOTA_CHANNEL=
-PLAYER_1_ID=
-PLAYER_2_ID=
-PLAYER_3_ID=
-PLAYER_4_ID=
-TWITCH_OAUTH_TOKEN=
-TWITCH_CLIENT_ID=
-TWITCH_USER_1=
-TWITCH_USER_2=
