+++ emoteLeaderboard.py
@@ -1,0 +1,146 @@
+from emoji import UNICODE_EMOJI
+from dotenv import load_dotenv
+import math
+import os
+import shelve
+import operator
+import re
+from collections import Counter
+
+load_dotenv()
+# Announcement
+ANNOUNCEMENT_CHANNEL = os.getenv('ANNOUNCEMENT_CHANNEL')
+ANNOUNCEMENT_DAY = int(os.getenv('ANNOUNCEMENT_DAY'))
+ANNOUNCEMENT_HOUR = int(os.getenv('ANNOUNCEMENT_HOUR'))
+ANNOUNCEMENT_MIN = int(os.getenv('ANNOUNCEMENT_MIN'))
+
+s = shelve.open('weekly_georgechamp_shelf.db')
+s_all_time = shelve.open('all_time_georgechamp_shelf.db')
+
+deleteMsg = None
+
+
+def is_emoji(s):
+    return s in UNICODE_EMOJI
+
+
+def score_algorithm(emoji_count):
+    return 0.61 + (1.37 * math.log(emoji_count))
+
+
+def updateCounts(key, increment=1):
+    try:
+        if s.get(key) is None:
+            s[key] = increment
+        else:
+            s[key] += increment
+
+        if s_all_time.get(key) is None:
+            s_all_time[key] = increment
+        else:
+            s_all_time[key] += increment
+
+        return True
+    except Exception:
+        return False
+
+
+async def announcement_task(channel):
+    shelf_as_dict = dict(s)
+    most_used_emotes = dict(sorted(shelf_as_dict.items(), key=operator.itemgetter(1), reverse=True)[:5])
+
+    keys = []
+    key_vals = []
+    for key in most_used_emotes.keys():
+        keys.append(key)
+        key_vals.append(most_used_emotes[key])
+
+    leaderboard_msg = "Weekly emote update: \nEmote - Score \n"
+    for i in range(5):
+        if i < len(keys):
+            leaderboard_msg = leaderboard_msg + str(i + 1) + ". " + keys[i] + " - " + str(key_vals[i]) + "\n"
+
+    global deleteMsg
+    if deleteMsg != None:
+        deleteMsg.delete()
+    deleteMsg = await channel.send(leaderboard_msg)
+
+    s.clear()
+
+
+async def print_count(message):
+    try:
+        requested_emote = message.content[10:]
+        await message.channel.send(requested_emote + " has been used " + str(s_all_time[requested_emote]) + " times.")
+    except KeyError:
+        await message.channel.send("Looks like that emote hasn't been used yet.")
+    except IndexError:
+        await message.channel.send("Doesn't appear that you've added an emote, please add an emote to check.")
+
+
+async def print_leaderboard(message):
+    await message.channel.send("Doesn't look like there are emojis here :( Try another page.")
+    shelf_as_dict = dict(s_all_time)
+    start = 0
+    end = 10
+
+    if len(message.content) != len('!leaderboard'):
+        increment = int(message.content[len('!leaderboard') + 1:])
+        # page size = 10
+        start += (increment - 1) * 10
+        end += (increment - 1) * 10
+
+    curr_page_num = (start / 10) + 1
+    total_page_num = len(dict(s_all_time).keys()) / 10 + 1
+    most_used_emotes = dict(sorted(shelf_as_dict.items(), key=operator.itemgetter(1), reverse=True)[start:end])
+    keys = []
+    key_vals = []
+    for key in most_used_emotes.keys():
+        keys.append(key)
+        key_vals.append(most_used_emotes[key])
+
+    if len(keys) == 0:
+        await message.channel.send("Doesn't look like there are emojis here :( Try another page.")
+    else:
+        leaderboard_msg = "All time leaderboard: - Page " + str(int(curr_page_num)) + "/" + str(
+            int(total_page_num)) + "\nEmote - Score \n"
+        for i in range(10):
+            if i < len(keys):
+                placement = start + i + 1
+                leaderboard_msg = leaderboard_msg + str(placement) + ". " + keys[i] + " - " + str(key_vals[i]) + "\n"
+
+        await message.channel.send(leaderboard_msg)
+
+
+async def check_emoji(message):
+    custom_emojis = re.findall(r'<:\w*:\d*>', message.content)
+
+    emoji_names = list(Counter(custom_emojis).keys())
+    emoji_counts = list(Counter(custom_emojis).values())
+    for i in range(len(emoji_names)):
+        updateCounts(emoji_names[i], round(score_algorithm(emoji_counts[i])))
+
+    unicode_emojis = []
+    for character in message.content:
+        if is_emoji(character):
+            unicode_emojis.append(character)
+
+    emoji_names = list(Counter(unicode_emojis).keys())
+    emoji_counts = list(Counter(unicode_emojis).values())
+
+    for i in range(len(emoji_names)):
+        updateCounts(emoji_names[i], round(score_algorithm(emoji_counts[i])))
+
+
+async def check_reaction(payload):
+    if payload.emoji.is_custom_emoji():
+        reaction_emoji_key = "<:" + payload.emoji.name + ":" + str(payload.emoji.id) + ">"
+        updateCounts(reaction_emoji_key)
+    elif payload.emoji.is_unicode_emoji():
+        updateCounts(payload.emoji.name)
+
+
+# s.close('weekly_georgechamp_shelf.db')
+# s_all_time.close('all_time_georgechamp_shelf.db')
+s.close()
+s_all_time.close()
